Change bookmark link case.
When generating markdown documents, it is generating bookmark link for each chapter on the top section. By the way, as those bookmark links had not converted to the lower case, it had occured mis-bookmark-linked problem when upper case being used in the chapter name.

This PR fixes such bug, by converting to the lower case alphabets for chaper bookmark links.