Optional cardinality in the 1 side.
When one-to-many relationship is optional in the one side, `prisma-markdown` could not express it in the diagram. Also, if one side is optional, but many side is mandatory, `prisma-markdown` also could not express it exactly, either.

This PR fixes the bug, so that one side optional cardinality can be exactly represented.